Update Readme.md
Added link to the Golang document. Made minor language updates. I also added a link to a medium article that describes what PKCE is. It was the only article I found that was not sponsored and the ietf specifications were complex worded. However, feel free to remove this link if you don't want it in the Readme.